벳스쿨 대기화면에러수정, 쇼핑몰삭제시 광고중인 상품이 있으면 삭제안되도록 처리

+++ wp-content/plugins/masterstudy-lms-learning-management-system/nuxy/metaboxes/metabox-shoppingmall-save.php
@@ -21,27 +21,34 @@ $newlink2 = $_POST['newlink2'];
 
 
 if($delcode){ // 삭제 요청
-    // 쇼핑몰리스트 활성화0 으로
-    $wpdb->get_results($wpdb->prepare("UPDATE wp_shoppingmall set state=0 where (code ='".$delcode."')"));
-
-    $result = $wpdb->get_results($wpdb->prepare("SELECT * FROM wp_product_list where (mall_code ='".$delcode."')"));
-    for($i=0; $i<count($result); $i++){
-        // 광고제품 리스트 광고활성화 0으로
-        $wpdb->get_results($wpdb->prepare("UPDATE wp_product_list set adv_state=0 where (ID ='".$result[$i] -> ID."')"));
-        // 영상에 등록되어져있는 제품들 삭제
-        $result2 = $wpdb->get_results($wpdb->prepare("SELECT * FROM wp_play_time where product_list_id =".$result[$i] -> ID));
-        for($j=0; $j<count($result2); $j++){
-            $wpdb->get_results($wpdb->prepare("DELETE FROM wp_play_time where (ID ='".$result2[$j]->ID."')"));
-        }
-        // 장바구니에 있는제품들 삭제
-        $result3 = $wpdb->get_results($wpdb->prepare("SELECT * FROM wp_wish_list where item_id =".$result[$i]->ID));
-        for($k=0; $k<count($result3); $k++){
-            $wpdb->get_results($wpdb->prepare("DELETE FROM wp_wish_list where (ID ='".$result3[$k]->ID."')"));
-        }
+    // 해당 쇼핑몰에 등록된 광고중인 상품이 있는지 확인
+    $sql = "SELECT * from wp_product_list where adv_state = 1 and mall_code =".$delcode;
+    $productcheck = $wpdb->get_results($wpdb->prepare($sql));
+    if($productcheck){ // 광고중인 상품있으면 삭제 막기
+        echo "광고중인 상품이 있습니다. 상품광고 비활성화 후 다시 시도해주세요.";
         
-
-
+    } else {
+        // 쇼핑몰리스트 활성화0 으로
+        $wpdb->get_results($wpdb->prepare("UPDATE wp_shoppingmall set state=0 where (code ='".$delcode."')"));
+    
+        $result = $wpdb->get_results($wpdb->prepare("SELECT * FROM wp_product_list where (mall_code ='".$delcode."')"));
+        for($i=0; $i<count($result); $i++){
+            // 광고제품 리스트 광고활성화 0으로
+            $wpdb->get_results($wpdb->prepare("UPDATE wp_product_list set adv_state=0 where (ID ='".$result[$i] -> ID."')"));
+            // 영상에 등록되어져있는 제품들 삭제
+            $result2 = $wpdb->get_results($wpdb->prepare("SELECT * FROM wp_play_time where product_list_id =".$result[$i] -> ID));
+            for($j=0; $j<count($result2); $j++){
+                $wpdb->get_results($wpdb->prepare("DELETE FROM wp_play_time where (ID ='".$result2[$j]->ID."')"));
+            }
+            // 장바구니에 있는제품들 삭제
+            $result3 = $wpdb->get_results($wpdb->prepare("SELECT * FROM wp_wish_list where item_id =".$result[$i]->ID));
+            for($k=0; $k<count($result3); $k++){
+                $wpdb->get_results($wpdb->prepare("DELETE FROM wp_wish_list where (ID ='".$result3[$k]->ID."')"));
+            }
+        }
     }
+
+
 
 
 }else if($editcode){ // 수정 사항 보여주기 요청
